Extract README parsing helpers to shared testutil package
Move duplicated README accuracy test helpers (FilePath, MarkdownLine,
TableCellName types and HCL/markdown parsing functions) from cnpg and
valkey test files to infra/testutil/readme_parser.go.

The shared package documents the README heading conventions relied upon:
- Output tables must appear under "## Outputs" (exact match)
- Input tables must appear under ## or ### headings containing
  "configuration" or "inputs" (case-insensitive)
- A new ## heading not containing "configuration" ends the inputs section

Both cnpg and valkey tests now import from testutil, eliminating ~200
lines of duplicated code per module.
