+++ infra/testutil/go.mod
@@ -2,7 +2,10 @@ module wildside/infra/testutil
 
 go 1.21
 
-require github.com/gruntwork-io/terratest v0.46.5
+require (
+	github.com/gruntwork-io/terratest v0.46.5
+	github.com/stretchr/testify v1.8.1
+)
 
 require (
 	cloud.google.com/go v0.110.0 // indirect
@@ -61,7 +64,6 @@ require (
 	github.com/pquerna/otp v1.2.0 // indirect
 	github.com/russross/blackfriday/v2 v2.1.0 // indirect
 	github.com/spf13/pflag v1.0.5 // indirect
-	github.com/stretchr/testify v1.8.1 // indirect
 	github.com/tmccombs/hcl2json v0.3.3 // indirect
 	github.com/ulikunitz/xz v0.5.10 // indirect
 	github.com/urfave/cli v1.22.2 // indirect
